#744 extracting string copy UI from user invite generator
